Renames files from `selector` to `select`
Since the `Selector` class is gone, it makes sense to name files as
`select` again.
