+++ src/frequenz/channels/util/_select.py
